WIP: select column component added. This component emits an array of selected columns.
